so close #2

+++ src/GameObject.java
@@ -1,30 +1,30 @@
 public abstract class GameObject {
-    protected int x;
-    protected int y;
+    protected double x;
+    protected double y;
 
     public GameObject() {
         this.x = 50;
         this.y = 50;
     }
 
-    public GameObject(int x, int y) {
+    public GameObject(double x, double y) {
         this.x = x;
         this.y = y;
     }
 
-    public void setX(int x) {
+    public void setX(double x) {
         this.x = x;
     }
 
-    public int getX() {
+    public double getX() {
         return x;
     }
 
-    public void setY(int y) {
+    public void setY(double y) {
         this.y = y;
     }
 
-    public int getY() {
+    public double getY() {
         return y;
     }
 }
