+++ src/Surface.java
@@ -1,31 +1,31 @@
 public class Surface {
     // Размерность карты
-    private int x;
-    private int y;
+    private double x;
+    private double y;
 
     public Surface() {
         this.x = 100;
         this.y = 100;
     }
 
-    public Surface(int x, int y) {
+    public Surface(double x, double y) {
         this.x = x;
         this.y = y;
     }
 
-    public int getX() {
+    public double getX() {
         return x;
     }
 
-    public int getY() {
+    public double getY() {
         return y;
     }
 
-    public void setX(int x) {
+    public void setX(double x) {
         this.x = x;
     }
 
-    public void setY(int y) {
+    public void setY(double y) {
         this.y = y;
     }
